data_pipeline/autoship_clean_function_source/requirements.txt
pandas
openpyxl
google-cloud-storage
google-cloud-bigquery
pyarrow